Fixed returning from `Read` if the timeout has expired
Update issue 90
Status: Review

There was a small bug in returning from `Read` keyword when the timeout has
expired. That is now fixed.

Also the keyword internals was changed so that `Read` always reads once
immediately and additionally, if `delay` has been given, the keyword reads
again at least once after the `delay`. Also the keyword documentation was
updated to match this behaviour.