Login keywords to read with delay if the prompt is not set
Update issue 90
Status: review

Keywords `Login` and `Login With Public Key` now have argument `delay`
to define how much to wait before reading from the server output again.
If the prompt is not set, this keyword reads and consumes the server
output similarly as with keyword `Read` with the given `delay`
(see issue 90 for details).

If the prompt is set, these keywords work as before.