Path separator is now depracted (and not removed)
Update issue 93
Labels: -bwic

Actually forgot that this needs to be deprecated first, so argument
`path_separator` now still works for `Get File` and `Put File` as before
but will trigger a deprecation warning. If the value is not given, the library
or the connection setting is used instead, as with the most of the other settings.